Bug Fix: Fix text exceeding screen boundaries
Until now, there was no type of check to verify that the location and
provider texts exceeded the maximum characters within the window.

This commit fixes this by checking the amount of characters (UTF-8)
that fit within the window and if it exceeds, it truncates the text
and adds an ellipsis to the end of the text.

Thanks to @equinoxmatt for reporting this =).
Issue #1